#594 [refactor] Separate CommentView
